requirements/base.txt
Django==2.0
django-environ>=0.4.0
django-braces>=1.8.1
django-crispy-forms>=1.5.2
django-admin-bootstrapped>=2.5.6
django-authtools>=1.4.0
easy-thumbnails>=2.3
django-material>=1.5.6
django-markdownx==2.0.28
django-haystack==2.6.1
elasticsearch>=2.0.0,<3.0.0
social-auth-app-django==2.1.0
